+++ internal/ent/task/task.go
@@ -74,7 +74,7 @@ func ValidColumn(column string) bool {
 
 var (
 	// TradingAccountIDValidator is a validator for the "trading_account_id" field. It is called by the builders before save.
-	TradingAccountIDValidator func(uint64) error
+	TradingAccountIDValidator func(int) error
 	// DefaultUpdatedAt holds the default value on creation for the "updated_at" field.
 	DefaultUpdatedAt func() time.Time
 	// UpdateDefaultUpdatedAt holds the default value on update for the "updated_at" field.
@@ -82,7 +82,7 @@ var (
 	// DefaultCreatedAt holds the default value on creation for the "created_at" field.
 	DefaultCreatedAt func() time.Time
 	// IDValidator is a validator for the "id" field. It is called by the builders before save.
-	IDValidator func(uint64) error
+	IDValidator func(int) error
 )
 
 // OrderOption defines the ordering options for the Task queries.
