+++ internal/ent/user/user.go
@@ -71,7 +71,7 @@ var (
 	// DefaultCreatedAt holds the default value on creation for the "created_at" field.
 	DefaultCreatedAt func() time.Time
 	// IDValidator is a validator for the "id" field. It is called by the builders before save.
-	IDValidator func(uint64) error
+	IDValidator func(int) error
 )
 
 // OrderOption defines the ordering options for the User queries.
